Added docker/check action

+++ dc-agent-app/src/main/java/com/payneteasy/dcagent/modules/docker/filesystem/FileSystemWriterImpl.java
@@ -2,6 +2,7 @@ package com.payneteasy.dcagent.modules.docker.filesystem;
 
 import com.payneteasy.dcagent.config.model.docker.Owner;
 import com.payneteasy.dcagent.modules.docker.IActionLogger;
+import com.payneteasy.dcagent.util.FileCompare;
 import com.payneteasy.dcagent.util.SafeFiles;
 
 import java.io.File;
@@ -12,6 +13,7 @@ import java.nio.file.attribute.PosixFilePermission;
 import java.util.HashSet;
 import java.util.Set;
 
+import static com.payneteasy.dcagent.util.FileCompare.isFileIdentical;
 import static com.payneteasy.dcagent.util.SafeFiles.*;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.nio.file.Files.copy;
@@ -26,6 +28,10 @@ public class FileSystemWriterImpl implements IFileSystem {
 
     @Override
     public void createDirectories(Owner aOwner, File aDir) {
+        if(aDir.exists()) {
+            return;
+        }
+
         logger.info("\uD83D\uDCC1 Creating directories {} ...", aDir.getAbsolutePath()); // 📁
         createDirs(aDir);
     }
@@ -34,12 +40,23 @@ public class FileSystemWriterImpl implements IFileSystem {
     public void writeExecutable(Owner aOwner, File aFile, String aText) {
         writeFile(aOwner, aFile, aText.getBytes(UTF_8));
 
-        logger.info("\uD83C\uDFBD Adding executable to {}", aFile.getAbsolutePath()); // 🎽
-
         Set<PosixFilePermission> perms = new HashSet<>();
         perms.add(PosixFilePermission.OWNER_READ);
         perms.add(PosixFilePermission.OWNER_WRITE);
         perms.add(PosixFilePermission.OWNER_EXECUTE);
+
+        Set<PosixFilePermission> existsPermissions;
+        try {
+            existsPermissions = Files.getPosixFilePermissions(aFile.toPath());
+        } catch (IOException e) {
+            throw new IllegalStateException("Cannot get attributes from file " + aFile.getAbsolutePath(), e);
+        }
+
+        if(perms.equals(existsPermissions)) {
+            return;
+        }
+
+        logger.info("\uD83C\uDFBD Adding executable to {}", aFile.getAbsolutePath()); // 🎽
 
         try {
             Files.setPosixFilePermissions(aFile.toPath(), perms);
@@ -65,6 +82,9 @@ public class FileSystemWriterImpl implements IFileSystem {
 
     @Override
     public void copyFile(Owner aOwner, File aFrom, File aTo) {
+        if(isFileIdentical(aFrom, aTo)) {
+            return;
+        }
         try {
             logger.info("\uD83D\uDDC3️ Copy file {} to {} ...", aFrom.getAbsoluteFile(), aTo.getAbsolutePath()); // 🗃️
 
@@ -80,6 +100,10 @@ public class FileSystemWriterImpl implements IFileSystem {
 
     @Override
     public void writeFile(Owner aOwner, File aSource, byte[] body) {
+        if(isFileIdentical(aSource, body)) {
+            return;
+        }
+
         logger.info("\uD83D\uDDC4️ Writing file {} ...", aSource.getAbsolutePath()); // 🗄️
         SafeFiles.writeFile(aSource, body);
     }
